+++ kotlin/coretypings/src/main/kotlin/eduflow/course/subscription/progress/CourseProgressLesson.kt
@@ -1,8 +1,8 @@
 package eduflow.course.subscription.progress
 
 interface CourseProgressLesson : CourseProgressItem {
-    val videoProgress: Int?
-    val synopsisProgress: Int?
+    val videoProgress: Double?
+    val synopsisProgress: Double?
     val isSurveyPassed: Boolean?
     // TODO: stuff
 }
